// model/customer-info-just-customer-id.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Conekta API
 * Conekta sdk
 *
 * The version of the OpenAPI document: 2.1.0
 * Contact: engineering@conekta.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface CustomerInfoJustCustomerId
 */
export interface CustomerInfoJustCustomerId {
    /**
     * 
     * @type {string}
     * @memberof CustomerInfoJustCustomerId
     */
    'customer_id': string;
}