// model/paymentMethodCash.ts
/**
 * Conekta API
 * Conekta sdk
 *
 * The version of the OpenAPI document: 2.1.0
 * Contact: engineering@conekta.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class PaymentMethodCash {
    'type'?: string;
    'object': string;
    'authCode'?: number | null;
    'cashierId'?: string | null;
    'reference'?: string;
    'barcodeUrl'?: string;
    'expiresAt'?: number;
    'serviceName'?: string;
    'store'?: string | null;
    'storeName'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "type",
            "baseName": "type",
            "type": "string"
        },
        {
            "name": "object",
            "baseName": "object",
            "type": "string"
        },
        {
            "name": "authCode",
            "baseName": "auth_code",
            "type": "number"
        },
        {
            "name": "cashierId",
            "baseName": "cashier_id",
            "type": "string"
        },
        {
            "name": "reference",
            "baseName": "reference",
            "type": "string"
        },
        {
            "name": "barcodeUrl",
            "baseName": "barcode_url",
            "type": "string"
        },
        {
            "name": "expiresAt",
            "baseName": "expires_at",
            "type": "number"
        },
        {
            "name": "serviceName",
            "baseName": "service_name",
            "type": "string"
        },
        {
            "name": "store",
            "baseName": "store",
            "type": "string"
        },
        {
            "name": "storeName",
            "baseName": "store_name",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return PaymentMethodCash.attributeTypeMap;
    }
}